requirements.txt
fastapi
uvicorn
tortoise-orm
gradio
requests